1.11.0

+++ src/main/java/com/github/hornta/race/objects/DurationUnit.java
@@ -1,17 +1,17 @@
 package com.github.hornta.race.objects;
 
 public class DurationUnit {
-  private final int amount;
+  private final long amount;
   private final String singularis;
   private final String pluralis;
 
-  public DurationUnit(int amount, String singularis, String pluralis) {
+  public DurationUnit(long amount, String singularis, String pluralis) {
     this.amount = amount;
     this.singularis = singularis;
     this.pluralis = pluralis;
   }
 
-  public int getAmount() {
+  public long getAmount() {
     return amount;
   }
